fix: adjusting component declarations and types
