+++ src/mocks/privileges/linixUseCases/LinixUseCases.mock.ts
@@ -3,7 +3,6 @@ export const linixUseCases = [
     id: "11",
     n_usecase: "Radicar Prospectos de crédito",
     k_usecase: "11212",
-    code: "11212",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12321312",
@@ -14,7 +13,6 @@ export const linixUseCases = [
     id: "12",
     n_usecase: "Viabilizar crédito",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Eliminar",
     a_publicc: "12112312",
@@ -25,7 +23,6 @@ export const linixUseCases = [
     id: "13",
     n_usecase: "Aprobar Créditos de alto monto",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -36,7 +33,6 @@ export const linixUseCases = [
     id: "14",
     n_usecase: "Aprobar Créditos de bajo monto",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Eliminar",
     a_publicc: "12112312",
@@ -47,7 +43,6 @@ export const linixUseCases = [
     id: "15",
     n_usecase: "Atender movimientos de caja",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -58,7 +53,6 @@ export const linixUseCases = [
     id: "16",
     n_usecase: "Registrar Movimientos",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -69,7 +63,6 @@ export const linixUseCases = [
     id: "17",
     n_usecase: "Registrar movimientos de pagos a proveedores",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -80,7 +73,6 @@ export const linixUseCases = [
     id: "18",
     n_usecase: "Contabilizar Ordenes de pago de Nomina",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -90,7 +82,6 @@ export const linixUseCases = [
     id: "19",
     n_usecase: "Registrar ajustes de activos fijos",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -100,7 +91,6 @@ export const linixUseCases = [
     id: "20",
     n_usecase: "Constituir Inversiones, Procesar la valorizacion",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Eliminar",
     a_publicc: "12112312",
@@ -110,7 +100,6 @@ export const linixUseCases = [
     id: "21",
     n_usecase: "Causar compras",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -120,7 +109,6 @@ export const linixUseCases = [
     id: "22",
     n_usecase: "Registrar Compras",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -130,7 +118,6 @@ export const linixUseCases = [
     id: "23",
     n_usecase: "Registrar movimientos de cartera",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -140,7 +127,6 @@ export const linixUseCases = [
     id: "24",
     n_usecase: "Configurar reglas de negocio",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -150,7 +136,6 @@ export const linixUseCases = [
     id: "25",
     n_usecase: "Hacer el cierre contable",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -160,7 +145,6 @@ export const linixUseCases = [
     id: "26",
     n_usecase: "Configurar reglas de negocio",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -170,7 +154,6 @@ export const linixUseCases = [
     id: "27",
     n_usecase: "Configurar reglas de negocio",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -180,7 +163,6 @@ export const linixUseCases = [
     id: "28",
     n_usecase: "Configurar reglas de negocio",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -190,7 +172,6 @@ export const linixUseCases = [
     id: "29",
     n_usecase: "Liquidar la Nómina",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Elimar",
     a_publicc: "12112312",
@@ -200,7 +181,6 @@ export const linixUseCases = [
     id: "30",
     n_usecase: "Contabilizar la Nómina",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -210,7 +190,6 @@ export const linixUseCases = [
     id: "31",
     n_usecase: "Registrar Incapacidades",
     k_usecase: "1231132",
-    code: "1231132",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -220,7 +199,6 @@ export const linixUseCases = [
     id: "32",
     n_usecase: "Registrar Vacaciones",
     k_usecase: "1231132",
-    code: "1231132",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -230,7 +208,6 @@ export const linixUseCases = [
     id: "33",
     n_usecase: "Registrar Licencias",
     k_usecase: "433232",
-    code: "433232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Elimar",
     a_publicc: "12112312",
@@ -240,7 +217,6 @@ export const linixUseCases = [
     id: "34",
     n_usecase: "Generar Orden de pago de Nómina",
     k_usecase: "1231132",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Eliminar",
     a_publicc: "12112312",
@@ -250,7 +226,6 @@ export const linixUseCases = [
     id: "35",
     n_usecase: "Liquidar Retiros de Empleados",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -261,7 +236,6 @@ export const linixUseCases = [
     id: "36",
     n_usecase: "Contabilizar Retiros de Empleados",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -272,7 +246,6 @@ export const linixUseCases = [
     id: "37",
     n_usecase: "Liquidar seguridad social",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -283,7 +256,6 @@ export const linixUseCases = [
     id: "38",
     n_usecase: "Contabilizar seguridad social",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -294,7 +266,6 @@ export const linixUseCases = [
     id: "39",
     n_usecase: "Registrar nuevos empleados",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Eliminar",
     a_publicc: "12112312",
@@ -304,7 +275,6 @@ export const linixUseCases = [
     id: "40",
     n_usecase: "Registrar Novedades de Retiro de empleados",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
@@ -314,7 +284,6 @@ export const linixUseCases = [
     id: "41",
     n_usecase: "Registrar Novedades de Seguridad Social",
     k_usecase: "131232",
-    code: "131232",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "21321",
@@ -324,7 +293,6 @@ export const linixUseCases = [
     id: "42",
     n_usecase: "Configurar reglamento de tasas",
     k_usecase: "523543",
-    code: "523543",
     n_descrip: "Explicacion prospectos de credito",
     i_tipusec: "Insertar",
     a_publicc: "12112312",
