+++ src/pages/privileges/outlets/linixUseCase/config/useCasesTable.config.tsx
@@ -3,7 +3,7 @@ import { Icon } from "@inube/design-system";
 import { Link } from "react-router-dom";
 import { linixUseCases } from "@src/mocks/privileges/linixUseCases/LinixUseCases.mock";
 import { DeleteUser } from "../../users/tabs/users/DeleteUser";
-import { DetailsModal } from "../details";
+import { DetailsModal } from "../components/details";
 
 const useCasesBreakPointsConfig = [
   { breakpoint: "(min-width: 1091px)", totalColumns: 4 },
@@ -42,14 +42,15 @@ const actionsConfig = [
   {
     id: "Details",
     actionName: "Detalles",
-    content: ({ id }: { id: string }) => (
-      <DetailsModal
-        icon={<MdOutlineAssignmentTurnedIn />}
-        infoDataModal={{ id }}
-        closeModal={() => {}}
-        linixUseCases={[]}
-      />
-    ),
+    content: ({ id }: { id: string }) => {
+      const useCase = linixUseCases.find((useCase) => useCase.id === id);
+      return useCase ? (
+        <DetailsModal
+          icon={<MdOutlineAssignmentTurnedIn />}
+          useCase={useCase}
+        />
+      ) : null;
+    },
     type: "secondary",
   },
   {
