+++ src/pages/privileges/outlets/linixUseCase/types.ts
@@ -1,0 +1,11 @@
+interface UseCase {
+  id: string;
+  k_usecase: string;
+  n_usecase: string;
+  n_descrip: string;
+  a_publicc: string;
+  i_tipusec: string;
+  n_camprv: string;
+}
+
+export type { UseCase };
